+++ src/io/reader/stdin.rs
@@ -1,6 +1,6 @@
 use crate::initial_read::InitialReadCompleteSender;
 use crate::io::controller::Reader;
-use crate::io::reader::utils::{read_bytes_until_newline, strip_newline_character};
+use crate::io::reader::utils::read_complete_lines;
 use crate::io::reader::{AsyncLineReader, ReadType};
 use async_trait::async_trait;
 use miette::Result;
@@ -23,7 +23,7 @@ impl StdinReader {
 #[async_trait]
 impl AsyncLineReader for StdinReader {
     async fn next(&mut self) -> Result<ReadType> {
-        let buffer = read_bytes_until_newline(&mut self.reader).await?;
+        let buffer = read_complete_lines(&mut self.reader).await?;
 
         if buffer.is_empty() {
             self.initial_read_complete_sender.send()?;
@@ -31,9 +31,6 @@ impl AsyncLineReader for StdinReader {
             return Ok(ReadType::StreamEnded);
         }
 
-        let buffer = strip_newline_character(buffer);
-        let line = String::from_utf8_lossy(&buffer).into_owned();
-
-        Ok(ReadType::SingleLine(line))
+        Ok(ReadType::MultipleLines(buffer))
     }
 }